explicitly casts alphaImage to a double type (e.g. in case it was a logical)

+++ packages/caffe_membrane_segmentation/Tests/overlay.m
@@ -10,5 +10,5 @@ hold on;
 h = imshow(green);
 hold off;
 
-set(h, 'AlphaData', alphaImage);
+set(h, 'AlphaData', double(alphaImage));
